Check for existing directories

+++ scripts/update_vim.sh
@@ -27,9 +27,13 @@ function update_vim() {
         sudo apt-get -y remove --purge vim vim-runtime vim-gnome vim-tiny vim-common vim-gui-common
         sudo apt-get -y build-dep vim-gnome
         sudo apt-get -y install liblua5.1-dev luajit libluajit-5.1 python-dev ruby-dev libperl-dev mercurial libncurses5-dev libgnome2-dev libgnomeui-dev libgtk2.0-dev libatk1.0-dev libbonoboui2-dev libcairo2-dev libx11-dev libxpm-dev libxt-dev
-        sudo mkdir /usr/include/lua5.1/include
-        sudo mv /usr/include/lua5.1/*.h /usr/include/lua5.1/include/
-        sudo ln -s /usr/bin/luajit-2.0.0-beta9 /usr/bin/luajit
+        if [[ ! -d /usr/include/lua5.1/include ]]; then
+            sudo mkdir -p /usr/include/lua5.1/include
+            sudo mv /usr/include/lua5.1/*.h /usr/include/lua5.1/include/
+        fi
+        if [[ ! -f /usr/bin/luajit ]]; then
+            sudo ln -s /usr/bin/luajit-2.0.0-beta9 /usr/bin/luajit
+        fi
         git clone https://github.com/vim/vim "$HOME/vim"
     fi
     cd "$HOME/vim/src"
